requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
